Refactor ExplicitResultTypes implementation
- remove dead code
- separate compiler API interactions to CompilerTypes helper
- move inner methods of the `fix` method into toplevel methods
